feat: update funcions and some bugfix

+++ Connection/ProductRepository.cs
@@ -25,9 +25,37 @@ public class ProductRepository : IProductRepository
     public void RemoveProduct(Guid productId)
     {
         var product = _context.Products.Find(productId);
-        _context.Products.Remove(product);
-        _context.SaveChanges();
-
+        
+        if(product != null)
+        {
+            _context.Products.Remove(product);
+            _context.SaveChanges();
+        }
     }
 
+    public void UpdateProduct(Guid productId, string? name = null, string? description = null, decimal price = -1)
+    {
+        var product = _context.Products.Find(productId);
+        if(product != null)
+        {   
+            if(name != null)
+            {
+                product.Name = name;
+            }
+            
+            if(description != null)
+            {
+                product.Description = description;
+            }
+
+            if(price > 0)
+            {
+                product.Price = price;
+            }
+
+            _context.Products.Update(product);
+            _context.SaveChanges();
+        }
+
+    }
 }
